Use rand instead of ring
`ring` contains everything but a kitchen sink when it comes to
cryptography. This crate doesn't need that, just randomness. `rand`
crate specializes for randomness only so it's smaller and quicker to
compile. It also doesn't need C compiler making deterministic builds
and cross-compiling easier.

The `RNGError` type was also modifed to make use of
`std::error::Error` which `rand::Error` implements as opposed to
`ring::error::Unspecified`.